Add --wait to iptables commands
This change adds the --wait flag to the insert/update/delete iptables
actions to prevent failures from occuring when iptables is running
outside of puppet. Previously if a script or a user is running iptables
commands while puppet is running, there may be failures in the firewall
provider if iptables cannot get a lock. By default, iptables will just
fail with an error so using --wait will make it wait until a lock can be
established. The --wait is only added for iptables >= 1.4.20.

Closes-Bug: #1612185
Change-Id: I4f667f61ef986f60441dbac3b440e383cc131451
Signed-off-by: Maksim Malchuk <mmalchuk@mirantis.com>